requirements.txt
streamlit
faster-whisper
google-generativeai
python-dotenv
requests
